refactor(viewer): extract shared refinement loop; de-hardcode demo-trace tests
The demand-driven refinement loop was copy-pasted across the
/api/flamegraph and /tokio-stats handlers and had diverged: tokio_stats
used cap = min(len, 100) while flamegraph used the real sampling_cap()
(5% fraction, baseline floor). Extract the algorithm into one deep module,
ingest/refine.rs (`refine() -> Option<Refined>`: list -> scope -> cap ->
fold -> coverage); both endpoints now call it and differ only in which
part-files they read and how they shape JSON.

This unification also:
- fixes the tokio_stats cap divergence (now the same policy as flamegraph)
- removes the polls_part_key_pub visibility shim (replaced by a proper
  aggregate::read_polls_parts)
- dedups the 4x AggContext construction into AppState::agg_context_for
- fixes tokio_stats' total_bytes: 0 coverage stub
- deletes the dead ordered_full_keys near-duplicate

Demo-trace test coupling: aggregate_test.rs hardcoded sample counts
(8971/8782/189/188/1) from the current demo trace, and the regen script
rewrote demo-trace.bin but not demo-trace.properties.json, so a regen
silently rotted both. Derive the expected counts from the fixture, and
have regenerate_demo_trace.sh refresh the fixture in the same step.

Sync CONTEXT.md (said baseline 3, cap 10%/300) and the aggregator docs to
the real code values (baseline 4, cap 5%/100) and the new refine.rs seam.

No trace-format changes. 114/114 nextest pass, clean under --stress-duration
20s, fmt + clippy clean, JS trace_properties self-test green.
